Split out sensors
