+++ custom_components/o365/classes/__init__.py
@@ -1,0 +1,1 @@
+"""Initialise."""
